Deleted file database.db Added files *.db to .gitignore Renamed file the database.py to db_utils.py Deleted some funcs in db_utils.py
