+++ .gitignore
@@ -55,6 +55,9 @@ cover/
 *.mo
 *.pot
 
+# DataBase
+*.db
+
 # Django stuff:
 *.log
 local_settings.py
